haru_cafe_django/requirements.txt
djangorestframework
pillow
psycopg
drf-spectacular
